refactor `#[derive(Visit)]` impl
